+++ src/net/ncplanner/plannerator/planner/s_tack/token/operator/LessOrEqualOperator.java
@@ -11,7 +11,7 @@ public class LessOrEqualOperator extends Operator{
     }
     @Override
     public StackObject evaluate(StackObject v1, StackObject v2){
-        return new StackBool(v1.asNumber().getValue().floatValue()<=v2.asNumber().getValue().floatValue());
+        return new StackBool(v1.asNumber().getValue().doubleValue()<=v2.asNumber().getValue().doubleValue());
     }
     
 }
