+++ src/net/ncplanner/plannerator/planner/s_tack/token/operator/PlusOperator.java
@@ -15,7 +15,7 @@ public class PlusOperator extends Operator{
         if(v1.getBaseType()==StackObject.Type.INT&&v2.getBaseType()==StackObject.Type.INT){
             return new StackInt(v1.asInt().getValue()+v2.asInt().getValue());
         }else{
-            return new StackFloat(v1.asNumber().getValue().floatValue()+v2.asNumber().getValue().floatValue());
+            return new StackFloat(v1.asNumber().getValue().doubleValue()+v2.asNumber().getValue().doubleValue());
         }
     }
 }
